Move addSchemaIndexFields to and created removeSchemaIndexFields in new file - this avoids bringing class files into the worker
